requirements.txt
streamlit==1.32.0
tensorflow==2.13.0
Pillow==9.5.0
keras==2.13.1
numpy==1.24.3
